feat: improve React loader and make it work with NextJS
