+++ packages/payment-widget/src/lib/react/global.d.ts
@@ -1,0 +1,8 @@
+declare namespace JSX {
+  interface IntrinsicElements {
+    "payment-widget": React.DetailedHTMLProps<
+      React.HTMLAttributes<HTMLElement>,
+      HTMLElement
+    >;
+  }
+}
