+++ packages/payment-widget/tsconfig.react.json
@@ -6,7 +6,10 @@
     "jsx": "react-jsx",
     "lib": ["ES2017", "DOM"],
     "esModuleInterop": true,
-    "allowSyntheticDefaultImports": true
+    "allowSyntheticDefaultImports": true,
+    "allowJs": true,
+    "noImplicitAny": false,
+    "strict": false
   },
   "include": ["src/lib/react/**/*"]
 }
